requirements.txt
imbalanced-learn==0.10.1
torchmetrics==0.11.4
ftfy==6.1.1
open-clip-torch==2.19.0
transformers==4.28.1
umap-learn==0.5.3
pandas==2.0.1
scikit-learn==1.2.2
nltk==3.8.1
psutil==5.9.5
urllib3==1.26.15
matplotlib==3.5.1
seaborn==0.11.2
h5py==3.8.0